[B] Move global settings to manage
